refactor: extract Hero into a page component
