+++ src/components/pages/home/index.js
@@ -1,0 +1,3 @@
+import Hero from './Hero'
+
+export { Hero }
